[BUGFIX release] Disable polymorphic deserialization when a model expects a type attribute
b7f7b7a introduced a bug where if there is a type key in a payload that is part
of an array, it would be used for polymorphic deserialization even when the
model expects an attribute that is named "type".

Interestingly, for such payloads, `arrayHash` passed into `normalizeArray()` in
rest-serializer.js contains Ember.Object instances as opposed to plain objects.
This causes the code to throw, since `hash.type` would be a computed property
in that case. This is probably because of extending from JSONAPISerializer.

(cherry picked from commit 69dca0c87cf6a830c4dee1214af4cb512f9806cb)